requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
openai
python-dotenv
PyPDF2
python-pptx